refactor: config.ts -> src/lib/supabase/supabaseConfig.ts로 이동
